+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 All notable changes to `laravel-nano-to` will be documented in this file
 
+## 1.0.2 - 2021-10-18
+
+- Use Nano.to POST Method as default checkout page generator
+- Ability to add additional Metadata when creating checkout page (Can be received back in Webhook)
+- Added Business (Name, Logo, Favicon), Background & Color customization
+- Ability to generate QR Code for RAW Nano (Supporting Natrium)
+- NanoToApi Advanced / Helper functions
+- PHP Unit Tests Updated & Added with multiple Issue Fixes
+
 ## 1.0.1 - 2021-09-19
 
 - Allow Custom Webhook Secret